Create structure1.c
In this program, we have created a struct named Person. We have also created a variable of Person named person1.
In main(), we have assigned values to the variables defined in Person for the person1 object.
strcpy(person1.name, "George Orwell");
person1.citNo = 1984;
person1. salary = 2500;
Notice that we have used strcpy() function to assign the value to person1.name.
This is because name is a char array (C-string) and we cannot use the assignment operator = with it after we have declared the string.